+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v1.1.0]() - 2022-09-23
+
+### Added
+- Create instance profile
+- Add EKS Role with policies
 
 ## [v1.0.0]() - 2022-07-07
 
 ### Added
-- DEVOPS-5138: Create module
+- Create module